Reorganised utils
