+++ formats/yarle/utils/turndown-rules/internal-links-rule.ts
@@ -1,4 +1,4 @@
-import { genUid } from '../../../../util';
+import { genUid } from "formats/keep/util";
 import { RuntimePropertiesSingleton } from '../../runtime-properties';
 import { yarleOptions } from '../../yarle';
